+++ pyproject.toml
@@ -1,0 +1,3 @@
+[tool.black]
+line-length = 79
+skip-string-normalization = true
